Tweaks to README
Thanks! I like this.

- fix typos
- be overly pedantic about colons
- emphasise the "one" command

I also think we should call the shell script `create-eks-cluster`, adding "eks" and removing the ".sh". 

It would be better if the output were cleaner and had less random debugging stuff, but I understand that might be a bit much to ask given the time constraints.