Ensure taskErr channel is buffered
The Problem:

Specifically under ClusterProvider.CreateCluster, there are 4
tasks that need to complete.

The first two tasks run in parallel. The next 2 run serially,
but are not gated against success of the first two.

This means that if more than one task fails, the cluster creation
process can get stuck attempting to write to the unbuffered
taskErr channel.

The operational fix:

Buffer the channel with 5 slots for errors, ensuring there
are no surprises.

eksctl is not some high-performance app, so having a buffer
is useful.

The systematic fix:

When writing to the channel, provide a default escape clause
that allows a failed write to be noticed, and warned of. This
will prevent silent hangings.