Make Delete idempotent and resumable
When EKS cluster is deleted, if the control plane does not
exist, the entire delete operation is aborted. This can
leave a lot of AWS artifacts lying around - Nodes, security groups,
subnets, VPCs, etc.

There is a real material cost to leaving these resources around.

This change makes delete a repeatable and idempotent operation.

It allows a paused/aborted/failed delete to be repeated ensuring
all resources are deleted eventually.