Revamp error channels
This change makes 2 major changes:
1. No action-at-a-distance when it comes to opening/closing channels

Channels are opened and defer-closed in the same scope, and only readable
channels are returned from functions.

All functions are non-blocking, returning immediately. All callers
are now forced to read only one channel to get any returned errors.

This also means channels are not accidentally abused, meaning
nobody can accidentally post to a read-only channel.

2. No more return-nil protocol.

By closing channels when their goroutines are complete,
any blocked readers on the channels will, by definition,
be unblocked instantly and they'll get the default
zero-value which is nil.

This removes another class of problems where all return paths
no longer need to be validated, so long as the channel-closing is
assured (which is easier since it's within the same scope.)