Don't import public key knowing it will conflict.
When re-running cluster create, the first step is importing a Public SSH Key
from the default location.

However, the fallback was counter-intuitive - if the file doesn't exist,
then an existing key (based on that filename) is searched for.

But if the file does exist and the key has been imported already,
the create command fails. But it doesn't really fail
in any helpful way, and makes scripting difficult.

This change flips the flow:
1. First, existence of the key is checked in EC2. If it exists, it is used.
2. If it does not exist, it is attempted to be imported.

This means if the create command is executed from a different host, the caller
has to ensure THEIR key is used correctly.