AMI resolution and overriding
If an AMI isn't specified then it will be resolved based
on region and instance type. It uses "resolvers" to do this.
Currently there are 2 resolvers:
- Default which uses the default EKS AMIs for a region
- GPU which for GPU instance types (p2 or p3) will use the
specific EKS GPU optimized image for a region.

The ami's used for non-gpu instances have been updatedt o use the
v23 version of the images. Also added the image for eu-west-1.

This change also allows you to specify an AMI to use instead of
using resolution.

If using the GPU instances then the [NVIDIA device plugin](https://github.com/NVIDIA/k8s-device-plugin)
will still need to be installed manually. Instructions are output. This
will be an 'addon' in the future.

Issue #183