Changed to have static and automatic static_resolvers
The code has been refactored so there are static resolvers that have
have the AMI image ids embedded into eksctl. There is also a 'auto'
resolver which will query AWS to get the AMI to use for the nodes.

The '--node-ami' flag can be used to supply an ami id to use. There are
2 special values that can be passed as well:
- 'fixed' which is the default and it will make sure the static
resolvers are used
- 'latest' which instructs eksctl to use the 'auto' resolver.

The static resolvers use maps to aid future go generate code which
will aid #49.