AMI resolution and overriding (allowing GPU support) (#192)
The `--node-ami` flag has been added and can be used to supply an ami id to use. There are
2 special values that can be passed as well:

- 'static' which is the default and it will make sure the static
resolvers are used (which uses embedded image ids)
- 'auto' which instructs eksctl to use the 'auto' resolver (which queries AWS).

If neither of these special values are supplied to the flag then we take it as a excplicit AMI image id that should be used for the EKS nodes.

The 'static' and 'auto' resolvers can resolve instance types with GPU support (p2/p3) and use an AMI image that is optimized with GPU support. If using the GPU instances then the [NVIDIA device plugin](https://github.com/NVIDIA/k8s-device-plugin) will still need to be installed manually. Instructions are output. This will be an 'addon' in the future.

The ami's used for non-gpu instances have been updated to use the v23 version of the images. Also added the image for eu-west-1.

The static resolvers use maps to aid future go generate code which will aid #49.

After changing the AMIs to v23 a change has been made to use aws-iam-authenticator instead of heptio-authenticator-aws.

Also, AMI resolution is done earlier before the stack is started to be built...fail fast. Also added code to check the resolved image exists.

Issue #183 
Issue #197 
Issue #194
Issue #169