Added eksctl get supported-regions command
In response to issue #222, I added a command to allow a user
to get the regions where EKS is currently supported. After
some discussion with another contributor, thought it would
be better under the "get" command.

The command is simple, and prints out a few lines showing
which regions EKS is currently supported in. I also added
the url where the most up to date information is on AWS.

Issue #222