Changed region handling
The region handling for the commands has changed. Previously, if
you didn't specify the region flag (-r/--region) then we would
default to us-west-2 even if you had a different region specified
in the AWS environment variables or shared config. With this change
we only default to us-west-2 if there is no region specified using the
flay AND there is no region coming from the AWS shared config /
environment variables.

Issue #235