Reworked region logic
Changed the the region logic so that if we create a AWS session
and there is no region from either the flags, aws shared config,
and aws environment variables then we re-create the session again
but specifying a default region. This is less than ideal but the
shared config isn't loaded until we create new session and the code
to read the shared config (ini files) is private.

A new issue will be opened to change this in the future to be cleaner
but we need a change to the AWS SDK first.

Signed-off-by: Richard Case <richard.case@outlook.com>