Retrieve IPv4 address from meta-data
Using meta-data is a reliable way to ensure that the NODE_IP is set to
the primary private IPv4 address instead of a randomly-selected secondary IP
(the AWS CNI will assign a number of secondary IPs to the node).

If the node is not registered with a primary address, the master will
be unable to communicate with nodes on port 10250, which breaks
functionality like log retrieval and port forwarding.