Add a way for users to provide bootstrap command
This should allow those who _know what their are doing_ to pass
a custom script to run inside of cloud-init. Anyone looking to
use this will need to defined a shell script that would resemble
`bootstrap.{al2,ubuntu}.sh`. This should cater for any of advanced
use-cases, where additional files must be created on EC2 instances
prior to boot or anything of that nature. Alternative approach is
to use a custom AMI, or propose an eksctl feature, but this is
meant as a stop-gap for testing things out.