Installing Flux now deletes pre-existing resources
... prior to creating them again.

Indeed, the user's first installation might have partially failed,
or the user may want to reinstall Flux for whatever reason, but the
current implementation fails on a 2nd installation.

This is because `CreateOrUpdate` is not idempotent and fails on the
more complex objects, with important values defaulted server-side
like `Service` objects' IP addresses.

An alternative approach would be to use `kubectl apply` but this
would introduce a dependency on `kubectl`.

Note that we do NOT delete Flux's private SSH key, so that the user
does not have to set up Flux's permissions again in his repository.