Use --cluster for non-cluster commands
Commands act on a principal resource which may be a cluster or not. Most
commands act in the context of a cluster and as such need to receive its
name.

In case of the prinicpal resource being a cluster (such as `eksctl get cluster`)
the flag name is `--name`. For all other commands, the cluster name is
being passed in as `--cluster`.