Remove outdated and generated version/release.go
The file `pkg/version/release.go` is generated when doing a release
through `tag-release.sh` and also committed into its release branch but
never merged back to master.

The one in master is from before we changed this process and can be
confusing for third parties vendoring and seeing the code with an
outdated tag.