Refactoring: move create cluster's CLI flags validation to Loader
... as this is where all validation happens.
`CreateClusterCmdParams#Validate()` is called in both:

- `validateWithConfigFile` and
- `validateWithoutConfigFile`

even though, currently, the validation logic is only relevant to the
"without config file" path, as this is how it logically should be.

The reason for grouping validation logic of `CreateClusterCmdParams`
in `CreateClusterCmdParams#Validate()` instead of in the validation
functions is that an object should generally be able to validate
itself, as opposed to "leaking" its field members to other code
for it to validate them.