Fargate: more consistent logging, only log 'creating'/'deleting' at info level for long-running operations
Creating and deleting profiles in an asynchronous operation, but given
that the async. operation itself is quick, in general, we should only
log about the fact it is in progress at debug level.
When the user wants to make this synchronous by passing `--wait`, then
we log about it at info level as it becomes a slower operation.