Added support for overridding the EKS Endpoint for the following commands: - `eksctl get iamidentitymapping` - `eksctl create iamidentitymapping` - `eksctl delete iamidentitymapping` - `eksctl create iamserviceaccount` - `eksctl delete iamserviceaccount` - `eksctl create nodegroup` - `eksctl delete nodegroup` - `eksctl drain nodegroup` - `eksctl delete cluster`
Updated the Kubeconfig class so that it accepts an override for the EKS Endpoint.

Refactored explicit arguments into a params object. I followed the pattern used in pkg/ctl/create/cluster.go.

Added an explicit --cluster-endpoint-override flag for defining the EKS Endpoint you'd like to use for the eksctl commands.

This change allows users with a private cluster to be able to use SSH tunnelling to access the cluster from their local machine via a specified port.

Issue #1552