No longer print usage/help on errors
This is to avoid printing usage/help on genuine runtime errors.
However, a side effect of this is that we now also no longer print
usage/help on input validation errors either.

That being said, given:

- we're trying to sanitise/simplify `eksctl`'s code at the moment,
- other CLI also do NOT print any usage/help on input validation errors
  e.g. `kubectl`,

it was decided between Martina, Chetan and I that it didn't matter too
much for now.