Identity removal from nodegroup deletion postponed to after.
The identity removal of the role ARN from the auth configmap was moved
to after the nodegroup deletion when the wait flag is used.
This was done to avoid issues where a nodegroup fails to delete for
whatever reason and be left with orphaned nodes without an identity
if it were to be deleted up front before the deletion of the nodegroup(s).

Deletion of identities from the auth configmap map via the delete
nodegroup call have been moved to tasks for more verbose planning.

Also for nodegroup deletion, stack descriptions have now been cached to
avoid duplicate calls to the Cloudformation API.

Issue #1703, #1368