Support for offline build and test
There are files used in the build/test to go generate that make
calls to retrieve files from online resources.
This change would allow using artifacts already downloaded.

Artifacts needed:
- pkg/nodebootstrap/maxpods.go
- pkg/addons/default/assets/aws-node.yaml

To update the files run:
make download-assets

If any of the artifacts are missing, then it will attempt to dowload
them when building.

Note that when submitting a PR, you should update the files to get
the latest files.

Issue #1608