Added kube-proxy version based on AWS Docs
Some kubernetes control plane versions don't have a corresponding kube-proxy image tag with the same version. This change uses major and minor version to lookup for a documented version if exists, otherwise it uses the control plane version to build the image tag. The only consideration is to make sure to update the mapping each time AWS releases a new documented version.

Issue #2536