Remove the pull_request_review trigger
It will always fail for forks and it's unlikely that a PR review
will be the step before an automerge (usually the label comes after
an approval OR a check completes).
In cases where it would have triggered the workflow and a merge,
the user can always manually merge after their approval anyway.