Handle aws-node 1.7 initContainer properly
Since version 1.7 of AWS VPC CNI plugin for Kubernetes
the DaemonSet consists additionally of one initContainer.

This initContainer needs the same treatment as the main container regarding
usage of the regional image.

Otherwise the consequence would be a failing pod in China regions
or slow image pulls in other regions than us-west-2.