Add Ubuntu 20.04 Image Support (#2758)
* Added Ubuntu 20.04 Image Support

Starting with EKS 1.18, the official Ubuntu EKS images are based on
Ubuntu 20.04 LTS (Focal). EKS 1.17 is the last release to use
Ubuntu 18.04 LTS (Bionic) as a base.

There are two key changes included:

* Add CLI support for the Ubuntu2004 family
* Create a new bootstrap script specific to Ubuntu2004

Ubuntu 20.04 EKS images, are no longer disabling the
systemd-resolved's DNS stub listener. Without this, the CoreDNS will
assume there is a loop when trying to incorrectly use
/etc/resolv.conf.

* Add automated tests to api for Ubuntu2004 family

* Improve amiFamily JSON schema

* Reduce duplication

* Make sure Ubuntu1804/2004 finds correct images

Co-authored-by: Michael Beaumont <mjboamail@gmail.com>