Add option to specify IAM role name
Currently, CloudFormation generates a role name that contains
a random suffix. However, in some use-cases a predetermined
role name is useful. For instance, when following a git-ops
workflow it might be necessary to include the serviceaccount
annotations for the IAM role ARNs in the manifest repo. With
random role names that can only be done after the role has
been created by eksctl. In addition, if the cluster should ever
get re-created all role ARNs would change which would then
require all manifests that reference those ARNs to be edited.

Closes #2261