Fixed issue where get-clusters prints null as opposed to []. Wrote a test for the same.
Issue: The printer for json and yaml type output will print "null" if the value supplied to it is nil and will print "[]" when the value supplied is an empty slice. The function responsible for returning the clusters output was sending nil when no clusters exist.

Fix: Intialized the variable to an empty slice in listClusters so that it returns an empty slice instead of nil. Wrote a test which will check if ListClusters returns the correct output i.e empty slice.

Issue #2864