Improve handling of kube-proxy upgrade
Starting from EKS 1.21 AWS has decided it will now include a the commit
sha in the kube version. This causes problems when upgrading kube-proxy
because it sets the image name to include build part of the semver so
image pulls fail.

Starting from 1.18 all kube-proxy images have a postfix of eksbuild.2.
Assume this is true for all release from 1.18 onwards