Implement dumpLogsToDisk function
Since we want to dump logs to disk only in the event of failure, we need to buffer the logs until a failure is encountered. This buffer (logBuffer) is provided as a writer to the logger library, and the output is mulit-written to this buffer and the original destination. For better organization purpose, the failure logs are stored in a newly created "logs/" directory, and are timestamped.