Dump logs to disk on failure (#4343)
* Implement dumpLogsToDisk function

Since we want to dump logs to disk only in the event of failure, we need to buffer the logs until a failure is encountered. This buffer (logBuffer) is provided as a writer to the logger library, and the output is mulit-written to this buffer and the original destination. For better organization purpose, the failure logs are stored in a newly created "logs/" directory, and are timestamped.

* Add debug log directory path to gitignore

The log directory and it's contents are created for debug purposes. No need for this to remain in the repo.

* Add actual error message to dumped file

* Dump logs to disk on failure integration test

A `eksctl get --name <does_not_exist>` command is run, with assertions.
These assertions check if the code exits with status 1, and if a new
failure log file is generated. The code also performs necessary cleanup.

* Improved error handling

- Inline handling of errors for consistency
- Using logger instead of fmt.Println for printing error info to stdout

* Create only one log file per day

- Instead of new file for each failure, have one log file per day
- This file will contain a timestamped version of the logs and error
   message for a particular day
- Reduces clutter in the generated files, and makes logs easier to find

* Make dump logs to disk opt-in

- Use -d flag to enable dump logs to disk during failure
- Updated integration test to reflect this behaviour

* Removing underscore from integration test package

Co-authored-by: Himangini <himangini@weave.works>
Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>