Handle pod eviction errors correctly (#5116)
* Handle pod eviction errors correctly

Currently any eviction error causes the draining of a node to stop and a new
node to start draining.  Eviction errors are common, expected occurences
especially when PDBs are used in the cluster.
By having any error abort the draining of a node we slow down the entire node
draining process as many of the pods further in the list could happily be
drained.

This change separates recoverable and irrecoverable eviction errors and retries
only the recoverable.  Unrecoverable errors fail the entire command.

An important aspect of this is that the `evictPods` function becomes blocking
until a node is drained or the process times out.  This is required as the
current implementation begins draining another node on the first eviction
error.  We would rather keep trying and eventually time out than make a bad
situation worse by draining a new node.

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Update pkg/drain/nodegroup.go

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>

* Use apimachinery error functions to check for recoverable errors

* Add --pod-eviction-wait-period flag to cluster delete command

Co-authored-by: William Hutcheson <william.hutcheson@skyscanner.net>
Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>