Parallel drain all nodegroups (#5146)
Currently all nodegroups are drained in serial but the nodes in each node group
are drained in parallel.  This change updates this logic to drain all nodes
across all node groups in parallel - up a maximum defined by the --parallel
flag.

This significantly speeds up draining for clusters with a large number of small
node groups.  Especially if some of the nodes in these node groups take longer
to drain than others.

This has the side affect of cordoning all nodes in all node groups before
beginning the actual draining process.  This should be the preferred behaviour
but may come as a surprise to some users - especially those that assume the
cordoning is actually the node being drained.

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>