Ignore IAM OIDC errors if no OIDC provider is associated with cluster
This changelist tags the cluster with the value of iam.withOIDC to allow eksctl to determine during cluster deletion if an OIDC provider was associated with the cluster or not.
Upon encountering an IAM permissions error, it throws an error if the cluster has an OIDC provider associated with it, which is determined either by the tag, or by the presence of IAM service accounts for clusters that lack the tag.