Append autoscaling tags with PropagateASGTags
The current behaviour is to only add the tags

 k8s.io/cluster-autoscaler/enabled = 'true'
 k8s.io/cluster-autoscaler/CLUSTER_NAME = 'owned'

to an ASG when WithAddonPolicies is enabled. However,
the IAM policy is only required for nodes that host
the autoscaling controller. Additionally, clusters that don't
rely on IAM instance profiles but use OIDC authentication
for the autoscaling controller don't need the IAM policy addon at
all.
Since we can assume that all nodegroups that have PropagateASGTags
enabled are supposed to be auto-scaled it makes sense to also add
the autoscaling tags to those ASGs.