Fix updating kube-proxy to a newer version
The version returned by `DescribeAddonVersions` cannot be used as-is for kube-proxy as it does not always map to the container image version. But for now, it does map to the minimal image tag.

This changelist adds a `minimal` string to the version obtained from `DescribeAddonVersions` to map it to a valid container image.