add comment about availabilityZones under managed node group
It is not currently clear to the user that a specific availability zone
can be requested under a managed node group. Having spent time / resources
and ultimately guessing about this param (and having looked on the page and
not seeing it) adding this to the documentation would be hugely useful.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>