Fix nesting of info text
The text here appears to belong inside the information bubble but due to nesting that is not how it is rendered.

What you see on the page currently is a collapsible info header that does nothing when clicked and this text, which causes confusion because of how it reads when not in the info bubble. Specifically, the part about "or upon cluster creation" tripped me up since I assume that is also when kube-proxy, aws-node, and coredns are also created; that made the info seem like it was indeed talking about the default addons.