cleanup efa installer archive before install
Currently, the UserData section that runs during cloud init
happens before any root volumes are expanded with growpart.
Although the best solution would be to ensure the filesystem
resize happens before these scripts are run, a quick means
to fix the current issue is simply to cleanup the efa
installer tar.gz, which is very large. I have tested this
with hpc7g for a size 2 and size 8 cluster (previously both
not working) and can confirm the devices are functioning after.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>